+++ podcasts/EpisodeFilterOverlayController.swift
@@ -1,4 +1,5 @@
 import UIKit
+import PocketCastsUtils
 
 class EpisodeFilterOverlayController: FilterSettingsOverlayController, UITableViewDataSource, UITableViewDelegate {
     static let episodeCellId = "CheckboxCellId"
@@ -31,6 +32,12 @@ class EpisodeFilterOverlayController: FilterSettingsOverlayController, UITableVi
 
         navigationController?.navigationBar.setValue(true, forKey: "hidesShadow")
         addCloseButton()
+
+        if FeatureFlag.playlistsRebranding.enabled {
+            handleThemeChanged()
+
+            saveButton.setTitle(L10n.playlistSmartRuleSaveButton, for: .normal)
+        }
     }
 
     // MARK: - TableView DataSource
@@ -63,12 +70,18 @@ class EpisodeFilterOverlayController: FilterSettingsOverlayController, UITableVi
         cell.episodeTitle.setLetterSpacing(-0.2)
         cell.selectButton.tag = tableRow.rawValue
         cell.selectButton.addTarget(self, action: #selector(selectButtonTapped), for: .touchUpInside)
-        cell.filterColor = filterToEdit.playlistColor()
+        if FeatureFlag.playlistsRebranding.enabled {
+            cell.episodeTitle.font = .systemFont(ofSize: 18, weight: .semibold)
+            cell.filterColor = AppTheme.colorForStyle(.primaryInteractive01)
+        } else {
+            cell.episodeTitle.font = .systemFont(ofSize: 16, weight: .medium)
+            cell.filterColor = filterToEdit.playlistColor()
+        }
         return cell
     }
 
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        51
+        FeatureFlag.playlistsRebranding.enabled ? 48 : 51
     }
 
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
@@ -93,10 +106,18 @@ class EpisodeFilterOverlayController: FilterSettingsOverlayController, UITableVi
 
         if filterFinished || filterUnplayed || filterPartiallyPlayed {
             saveButton.isEnabled = true
-            saveButton.backgroundColor = filterToEdit.playlistColor()
+            if FeatureFlag.playlistsRebranding.enabled {
+                saveButton.alpha = 1
+            } else {
+                saveButton.backgroundColor = filterToEdit.playlistColor()
+            }
         } else {
             saveButton.isEnabled = false
-            saveButton.backgroundColor = AppTheme.disabledButtonColor()
+            if FeatureFlag.playlistsRebranding.enabled {
+                saveButton.alpha = 0.4
+            } else {
+                saveButton.backgroundColor = AppTheme.disabledButtonColor()
+            }
         }
         tableView.reloadData()
     }
@@ -113,5 +134,14 @@ class EpisodeFilterOverlayController: FilterSettingsOverlayController, UITableVi
         filterToEdit.filterPartiallyPlayed = filterPartiallyPlayed
 
         super.saveFilter()
+    }
+
+    override func handleThemeChanged() {
+        super.handleThemeChanged()
+
+        if FeatureFlag.playlistsRebranding.enabled {
+            saveButton.backgroundColor = AppTheme.colorForStyle(.primaryInteractive01)
+            changeNavTint(titleColor: nil, iconsColor: AppTheme.colorForStyle(.primaryIcon03))
+        }
     }
 }