+++ podcasts/PlaybackManager.swift
@@ -74,6 +74,9 @@ class PlaybackManager: ServerPlaybackDelegate {
 
     private(set) var transcriptsAvailable = false
 
+    /// The time the episode was last switched as tracked by handleCurrentlyPlayingEpisodeUpdated
+    private var episodeSwitchTime: Date?
+
     init() {
         queue = PlaybackQueue()
         queue.loadPersistedQueue()
@@ -93,6 +96,7 @@ class PlaybackManager: ServerPlaybackDelegate {
         NotificationCenter.default.addObserver(self, selector: #selector(refreshRemoteCommands), name: Constants.Notifications.remoteCommandSettingsChanged, object: nil)
         NotificationCenter.default.addObserver(self, selector: #selector(updateNowPlayingInfo), name: Constants.Notifications.userEpisodeUpdated, object: nil)
         NotificationCenter.default.addObserver(self, selector: #selector(updateAllNowPlayingData), name: .episodeEmbeddedArtworkLoaded, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(handleCurrentlyPlayingEpisodeUpdated), name: Constants.Notifications.currentlyPlayingEpisodeUpdated, object: nil)
 
         // run these on a background queue because some of them might call our singleton instance back, causing a crash because PlaybackManager.shared is called from the init method
         DispatchQueue.global().async {
@@ -1731,7 +1735,17 @@ class PlaybackManager: ServerPlaybackDelegate {
                     if Settings.legacyBluetoothModeEnabled(), seekEvent.positionTime < 1 {
                         FileLog.shared.addMessage("Remote control: ignoring changePlaybackPositionCommand, it's to 0 and legacy bluetooth mode is on")
                     } else {
-                        FileLog.shared.addMessage("Remote control: changePlaybackPositionCommand")
+                        FileLog.shared.addMessage("Remote control: changePlaybackPositionCommand to \(seekEvent.positionTime)")
+
+                        if FeatureFlag.limitPlaybackPositionChanges.enabled {
+                            // Check if we're still in the limiting window after an episode change
+                            if let switchTime = episodeSwitchTime,
+                               Date.now.timeIntervalSince(switchTime) < 2.0 {
+                                FileLog.shared.addMessage("Remote control: ignoring changePlaybackPositionCommand due to recent episode switch")
+                                return .commandFailed
+                            }
+                        }
+
                         seekTo(time: seekEvent.positionTime)
                     }
 
@@ -2053,6 +2067,13 @@ class PlaybackManager: ServerPlaybackDelegate {
 
         // update the cached copy of the now playing episode so we have the latest version of it
         queue.nowPlayingEpisodeChanged()
+    }
+
+    @objc private func handleCurrentlyPlayingEpisodeUpdated() {
+        // Update episode switch time when the currently playing episode changes
+        if FeatureFlag.limitPlaybackPositionChanges.enabled {
+            episodeSwitchTime = Date()
+        }
     }
 
     // MARK: - Interruptions