+++ podcasts/ReleaseDateFilterOverlayController.swift
@@ -1,4 +1,5 @@
 import UIKit
+import PocketCastsUtils
 
 enum ReleaseDateFilterOption: Int32, AnalyticsDescribable {
     case anytime = 0
@@ -63,6 +64,13 @@ class ReleaseDateFilterOverlayController: FilterSettingsOverlayController, UITab
 
         setCurrentReleaseDate()
         setupLargeTitle()
+
+        if FeatureFlag.playlistsRebranding.enabled {
+            handleThemeChanged()
+
+            saveButton.setTitle(L10n.playlistSmartRuleSaveButton, for: .normal)
+        }
+
         title = L10n.filterReleaseDate
         tableView.contentInsetAdjustmentBehavior = .never
 
@@ -87,8 +95,13 @@ class ReleaseDateFilterOverlayController: FilterSettingsOverlayController, UITab
         cell.title.setLetterSpacing(-0.2)
         cell.style = .primaryUi01
         cell.setSelectState(selectedIndex == indexPath.row)
-        let filterTintColor = filterToEdit.playlistColor()
-        cell.setTintColor(color: filterTintColor)
+        if FeatureFlag.playlistsRebranding.enabled {
+            cell.title.font = .systemFont(ofSize: 17, weight: .semibold)
+            cell.setTintColor(color: AppTheme.colorForStyle(.primaryInteractive01))
+        } else {
+            cell.title.font = .systemFont(ofSize: 16, weight: .medium)
+            cell.setTintColor(color: filterToEdit.playlistColor())
+        }
         cell.selectButton.tag = indexPath.row
         cell.selectButton.addTarget(self, action: #selector(selectButtonTapped), for: .touchUpInside)
         return cell
@@ -100,7 +113,7 @@ class ReleaseDateFilterOverlayController: FilterSettingsOverlayController, UITab
     }
 
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        51
+        FeatureFlag.playlistsRebranding.enabled ? 46 : 51
     }
 
     // MARK: - Helper functions
@@ -124,5 +137,14 @@ class ReleaseDateFilterOverlayController: FilterSettingsOverlayController, UITab
 
         selectedIndex = buttonTag
         tableView.reloadData()
+    }
+
+    override func handleThemeChanged() {
+        super.handleThemeChanged()
+
+        if FeatureFlag.playlistsRebranding.enabled {
+            saveButton.backgroundColor = AppTheme.colorForStyle(.primaryInteractive01)
+            changeNavTint(titleColor: nil, iconsColor: AppTheme.colorForStyle(.primaryIcon03))
+        }
     }
 }